Re-factor upload helper
