+++ js/ui/file_uploader.js
@@ -1,4 +1,4 @@
-import { getWidth } from '../core/utils/size';
+import { getOffset, getWidth } from '../core/utils/size';
 import $ from '../core/renderer';
 import Guid from '../core/guid';
 import { getWindow } from '../core/utils/window';
@@ -15,11 +15,11 @@ import Editor from './editor/editor';
 import Button from './button';
 import ProgressBar from './progress_bar';
 import devices from '../core/devices';
-import { addNamespace } from '../events/utils/index';
+import { addNamespace, isTouchEvent } from '../events/utils/index';
 import { name as clickEventName } from '../events/click';
 import messageLocalization from '../localization/message';
 import { isMaterial } from './themes';
-import { isMouseOverElement, reRaiseEvent } from './file_uploader_helper';
+import domAdapter from '../core/dom_adapter';
 
 // STYLE fileUploader
 
@@ -923,7 +923,7 @@ class FileUploader extends Editor {
         }
 
         const dropZoneElement = this._getDropZoneElement(isCustomTarget);
-        if(dropZoneElement === e.target && this._activeDropZone === null && isMouseOverElement(e, dropZoneElement, false)) {
+        if(dropZoneElement === e.target && this._activeDropZone === null && this.isMouseOverElement(e, dropZoneElement, false)) {
             this._activeDropZone = dropZoneElement;
             this._tryToggleDropZoneActive(true, isCustomTarget, e);
         }
@@ -936,11 +936,11 @@ class FileUploader extends Editor {
         e.originalEvent.dataTransfer.dropEffect = 'copy';
         const dropZoneElement = this._getDropZoneElement(isCustomTarget);
 
-        if(this._activeDropZone === null && isMouseOverElement(e, dropZoneElement, false)) {
-            reRaiseEvent(e, 'dragenter', dropZoneElement);
+        if(this._activeDropZone === null && this.isMouseOverElement(e, dropZoneElement, false)) {
+            this._reRaiseEvent(e, 'dragenter', dropZoneElement);
         }
         if(this._activeDropZone !== null && this._shouldRaiseDragLeave(e, isCustomTarget)) {
-            reRaiseEvent(e, 'dragleave', this._activeDropZone);
+            this._reRaiseEvent(e, 'dragleave', this._activeDropZone);
         }
     }
 
@@ -954,7 +954,7 @@ class FileUploader extends Editor {
 
         if(this._shouldRaiseDragLeave(e, isCustomTarget)) {
             if(this._activeDropZone !== e.target) {
-                reRaiseEvent(e, 'dragleave', this._activeDropZone);
+                this._reRaiseEvent(e, 'dragleave', this._activeDropZone);
             } else {
                 this._tryToggleDropZoneActive(false, isCustomTarget, e);
                 this._activeDropZone = null;
@@ -963,7 +963,7 @@ class FileUploader extends Editor {
     }
 
     _shouldRaiseDragLeave(e, isCustomTarget) {
-        return !isMouseOverElement(e, this._activeDropZone, !isCustomTarget);
+        return !this.isMouseOverElement(e, this._activeDropZone, !isCustomTarget);
     }
 
     _tryToggleDropZoneActive(active, isCustom, event) {
@@ -1220,6 +1220,96 @@ class FileUploader extends Editor {
         this._totalFilesSize = 0;
         this._totalLoadedFilesSize = 0;
         this._updateTotalProgress(this._getTotalFilesSize(), this._getTotalLoadedFilesSize());
+    }
+
+    isMouseOverElement(mouseEvent, element, correctPseudoElements) {
+        if(!element) return false;
+
+        const beforeHeight = correctPseudoElements ? parseFloat(window.getComputedStyle(element, ':before').height) : 0;
+        const afterHeight = correctPseudoElements ? parseFloat(window.getComputedStyle(element, ':after').height) : 0;
+        const x = getOffset(element).left;
+        const y = getOffset(element).top + beforeHeight;
+        const w = element.offsetWidth;
+        const h = element.offsetHeight - beforeHeight - afterHeight;
+        const eventX = this._getEventX(mouseEvent);
+        const eventY = this._getEventY(mouseEvent);
+
+        return eventX >= x && eventX < (x + w) && eventY >= y && eventY < (y + h);
+    }
+    _reRaiseEvent(e, eventType, newTarget) {
+        const document = domAdapter.getDocument();
+        let newEvent;
+        if(window.MouseEvent) {
+            newEvent = new window.MouseEvent(eventType, {
+                bubbles: e.bubbles,
+                cancelable: e.cancelable,
+                window: e.window,
+                detail: e.detail,
+                screenX: e.screenX,
+                screenY: e.screenY,
+                clientX: e.clientX,
+                clientY: e.clientY,
+                ctrlKey: e.ctrlKey,
+                altKey: e.altKey,
+                shiftKey: e.shiftKey,
+                metaKey: e.metaKey,
+                button: e.button,
+                relatedTarget: e.relatedTarget
+            });
+
+            newTarget.dispatchEvent(newEvent);
+        } else if(document.createEventObject) {
+            newEvent = document.createEventObject(window.event);
+            newEvent.type = eventType;
+            newEvent.bubbles = e.bubbles;
+            newEvent.cancelable = e.cancelable;
+            newEvent.view = window;
+            newEvent.detail = e.detail;
+            newEvent.screenX = e.screenX;
+            newEvent.screenY = e.screenY;
+            newEvent.clientX = e.clientX;
+            newEvent.clientY = e.clientY;
+            newEvent.ctrlKey = e.ctrlKey;
+            newEvent.altKey = e.altKey;
+            newEvent.shiftKey = e.shiftKey;
+            newEvent.metaKey = e.metaKey;
+            newEvent.button = e.button;
+            newEvent.relatedTarget = e.relatedTarget;
+
+            newTarget.fireEvent(eventType, newEvent);
+        }
+    }
+    _getEventX(e) {
+        return isTouchEvent(e) ? this._getTouchEventX(e) : e.clientX + this._getDocumentScrollLeft();
+    }
+    _getEventY(e) {
+        return isTouchEvent(e) ? this._getTouchEventY(e) : e.clientY + this._getDocumentScrollTop();
+    }
+    _getTouchEventX(e) {
+        let touchPoint = null;
+        if(e.changedTouches.length > 0) {
+            touchPoint = e.changedTouches;
+        } else if(e.targetTouches.length > 0) {
+            touchPoint = e.targetTouches;
+        }
+        return touchPoint ? touchPoint[0].pageX : 0;
+    }
+    _getTouchEventY(e) {
+        let touchPoint = null;
+        if(e.changedTouches.length > 0) {
+            touchPoint = e.changedTouches;
+        } else if(e.targetTouches.length > 0) {
+            touchPoint = e.targetTouches;
+        }
+        return touchPoint ? touchPoint[0].pageY : 0;
+    }
+    _getDocumentScrollTop() {
+        const document = domAdapter.getDocument();
+        return document.documentElement.scrollTop || document.body.scrollTop;
+    }
+    _getDocumentScrollLeft() {
+        const document = domAdapter.getDocument();
+        return document.documentElement.scrollLeft || document.body.scrollLeft;
     }
 
     _updateReadOnlyState() {
